More convenient response from the HAL client.
We now return a response that is basically a HAL document where every
relation (that the user asked for) is embedded.

This is based on the assumption that objects in _embedded are identical
to what you'd get by asking for the link. This isn't something the HAL
spec demands, but we do adhere to it in the HDO API.